508965 rounding up obligation calculation and changing type for Tonnage column and property (#183)
* 508965 - Rounds up obligation calculation result and changes tonne property type from float to integer

* 508965 - Rounding up obligation calculation result - migration script updated

* 508965 - Rounding up  - Added FC material back in

* 508965 - rounding up obligation calculation - Use Math.Ceiling rather than ...ToPositiveInfinity

+++ src/EPR.PRN.Backend.Obligation/Models/SubmissionCalculationRequest.cs
@@ -7,5 +7,5 @@ public class SubmissionCalculationRequest
     public Guid SubmissionId { get; set; }
     public string SubmissionPeriod { get; set; }
     public string PackagingMaterial { get; set; }
-    public double PackagingMaterialWeight { get; set; }
+    public int PackagingMaterialWeight { get; set; }
 }
